+++ apps/validator/src/main/scala/com/daml/network/validator/ValidatorApp.scala
@@ -660,6 +660,7 @@ class ValidatorApp(
               validatorTopupConfig,
               config.walletSweep,
               config.autoAcceptTransfers,
+              config.supportsSoftDomainMigrationPoc,
             )
           )
         else {