+++ apps/wallet/src/main/scala/com/daml/network/wallet/UserWalletService.scala
@@ -45,6 +45,7 @@ class UserWalletService(
     validatorTopupConfigO: Option[ValidatorTopupConfig],
     walletSweep: Option[WalletSweepConfig],
     autoAcceptTransfers: Option[AutoAcceptTransfersConfig],
+    supportsSoftDomainMigrationPoc: Boolean,
 )(implicit
     ec: ExecutionContext,
     mat: Materializer,
@@ -74,6 +75,7 @@ class UserWalletService(
       PackageIdResolver.inferFromAmuletRules(clock, scanConnection, loggerFactory),
     ),
     treasuryConfig,
+    supportsSoftDomainMigrationPoc,
     clock,
     store,
     walletManager,
@@ -88,6 +90,7 @@ class UserWalletService(
     ledgerClient,
     scanConnection.getAmuletRulesDomain,
     automationConfig,
+    supportsSoftDomainMigrationPoc,
     clock,
     domainTimeSync,
     domainUnpausedSync,